+++ packages/graphile-build-pg/src/plugins/PgBasicsPlugin.ts
@@ -1038,6 +1038,7 @@ export default (function PgBasicsPlugin(
     pgIgnoreIndexes = true, // TODO:v5: change this to false
     pgHideIndexWarnings = false,
     pgLegacyJsonUuid = false, // TODO:v5: remove this
+    pgAugmentIntrospectionResults,
   }
 ) {
   let pgOmit = baseOmit;
@@ -1071,6 +1072,7 @@ export default (function PgBasicsPlugin(
         pgField,
         sqlCommentByAddingTags,
         pgPrepareAndRun,
+        pgAugmentIntrospectionResults,
       };
       return build.extend(
         build,